software/pi-software/requirements.txt
paho_mqtt==1.6.1
picamera==1.13
